Fixed mpu9250 43 second bug and other sensor class bugs (#22)
* fixed mpu9250 43 second bug and other sensor class bugs

* formatting

* formatting

+++ Sensor.cpp
@@ -9,7 +9,7 @@
  * @param setup - t_datasetup - How many datasets are there? What are their units?
  * @param delta - uint16_t - The minimum time (in ms) between sensor read attempts.
  **/
-Sensor::Sensor(sensors_t sensor, arduino_t arduino, t_datasetup setup, uint16_t delta){
+Sensor::Sensor(sensors_t sensor, arduino_t arduino, t_datasetup setup, unsigned long delta){
     // Set up cache
     this->sensor = sensor;
     this->arduino = arduino;
